Interactive mode Readme, Clean up, Seperation
